Move ui utils to Utils
