+++ FlipperZero/UI/Sources/Common/Background.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+extension View {
+    var systemBackground: Color {
+        .init(UIColor.systemBackground)
+    }
+}
