+++ FlipperZero/UI/Sources/Utils/ApplicationSettings.swift
