+++ FlipperZero/UI/Sources/Utils/FaceID.swift
@@ -12,13 +12,3 @@ extension UIDevice {
         return false
     }
 }
-
-extension View {
-    func resignFirstResponder() {
-        UIApplication.shared.sendAction(
-            #selector(UIResponder.resignFirstResponder),
-            to: nil,
-            from: nil,
-            for: nil)
-    }
-}
