+++ FlipperZero/UI/Sources/Utils/FirstResponder.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+extension View {
+    func resignFirstResponder() {
+        UIApplication.shared.sendAction(
+            #selector(UIResponder.resignFirstResponder),
+            to: nil,
+            from: nil,
+            for: nil)
+    }
+}
