+++ FlipperZero/UI/Sources/Utils/Platform.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+extension View {
+    var iOS14: Bool {
+        ProcessInfo.processInfo.operatingSystemVersion.majorVersion == 14
+    }
+
+    var onMac: Bool {
+        ProcessInfo.processInfo.isiOSAppOnMac
+    }
+}
